Customer handling is now done in CustomerController.java.
You can now search for a specific customer, view their bookings and delete the customer in the search-customer endpoint.

Deleting a customer now deletes the images assigned to their bookings and deletes sets the bookings customer to null if the booking is in the past. If it is in the future, it cancels the appointment.

+++ src/main/java/com/example/demo/services/BookingService.java
@@ -16,6 +16,7 @@ import java.io.IOException;
 import java.time.LocalDate;
 import java.time.LocalDateTime;
 import java.time.LocalTime;
+import java.util.ArrayList;
 import java.util.List;
 import java.util.UUID;
 
@@ -60,14 +61,45 @@ public class BookingService {
     @Transactional
     public void deleteBooking(Booking booking){
         setBookableHoursRelatedToBookingToAvailable(booking);
-        s3ImageService.deleteImage(booking.getTattooImage().getUrl());
-        tattooImageService.deleteTattooImage(booking.getTattooImage());
+        TattooImage tattooImage = booking.getTattooImage();
+
+        if(tattooImage != null && tattooImage.getUrl() != null){
+            s3ImageService.deleteImage(booking.getTattooImage().getUrl());
+            tattooImageService.deleteTattooImage(tattooImage);
+        }
+
         bookingRepo.delete(booking);
+    }
+
+    public void deleteFutureBookingsAndSetPastBookingsCustomerToNull(List<Booking> bookings){
+        List<Booking> bookingsToSetCustomerToNull = new ArrayList<>();
+        List<Booking> bookingsToDelete = new ArrayList<>();
+
+        for(Booking booking: bookings){
+            if(booking.getDate().isBefore(LocalDateTime.now())){
+                s3ImageService.deleteImage(booking.getTattooImage().getUrl());
+                tattooImageService.deleteTattooImage(booking.getTattooImage());
+                booking.setCustomer(null);
+                bookingsToSetCustomerToNull.add(booking);
+            }else{
+                bookingsToDelete.add(booking);
+                setBookableHoursRelatedToBookingToAvailable(booking);
+            }
+        }
+
+        if(!bookingsToSetCustomerToNull.isEmpty()){
+            saveListOfBookings(bookingsToSetCustomerToNull);
+        }
+
+        if(!bookingsToDelete.isEmpty()){
+            deleteBookings(bookingsToDelete);
+        }
     }
 
     private void setBookableHoursRelatedToBookingToAvailable(Booking booking){
         BookableDate bookableDate = bookableDateService.getBookableDateByDate(LocalDate.from(booking.getDate()));
         List<BookableHour> bookableHoursOnDate = bookableDate.getBookableHours();
+
         for(BookableHour bookableHour: bookableHoursOnDate){
             LocalTime hour = bookableHour.getHour();
             LocalTime startTime = booking.getDate().toLocalTime();
@@ -79,13 +111,22 @@ public class BookingService {
                 }
             }
         }
+
         bookableDateService.saveBookableDate(bookableDate);
     }
 
     public void deleteBookings(List<Booking> bookings){
+        List<TattooImage> tattooImages = new ArrayList<>();
         for(Booking booking: bookings){
+            tattooImages.add(booking.getTattooImage());
+            s3ImageService.deleteImage(booking.getTattooImage().getUrl());
             setBookableHoursRelatedToBookingToAvailable(booking);
         }
+
+        if(!tattooImages.isEmpty()){
+            tattooImageService.deleteListOfTattooImages(tattooImages);
+        }
+
         bookingRepo.deleteAll(bookings);
     }
